Support using alternative rustls backends like aws-lc-rs
rustls has support for alternate backends, with built-in support for
aws-lc-rs, as well as support for custom user-provided backends.

Add a default-enabled feature for the default ring backend, and a
non-default feature for the aws-lc-rs backend. Also support building
without either backend, in which case we require the user to supply
their own `ClientConfig` and don't provide functions that assume a
default.